+++ src/layout/_types/type.ts
@@ -4,5 +4,5 @@ export interface sidebarListType {
 	title: string;
 	href: string;
 	svgName: keyof typeof svgObj;
-	onToggleSidebar: () => void;
+	onToggleSidebar?: () => void;
 }
